not streamlit, but created setup file, reorganised general dir and ran clean files
